+++ extras/tests/pc_based/test_02.cpp
@@ -139,6 +139,14 @@ class FastAccelStepperTest {
       assert((i == 0) || (old_planned_time_in_buffer > 0.005));
       old_planned_time_in_buffer = planned_time;
     }
+	// Empty the queue 
+      while (!s.isQueueEmpty()) {
+        rc.check_section(
+            &fas_queue_A.entry[fas_queue[0].read_idx & QUEUE_LEN_MASK]);
+        fas_queue[0].read_idx++;
+        fprintf(gp_file, "%.6f %.2f %d\n", rc.total_ticks / 1000000.0,
+                16000000.0 / rc.last_dt, rc.last_dt);
+      }
     fprintf(gp_file, "EOF\n");
     fprintf(gp_file, "plot $data using 1:2 with linespoints\n");
     fprintf(gp_file, "pause -1\n");
@@ -192,6 +200,7 @@ int main() {
 
   test.with_empty_queue();
   //             steps  ticks_us  accel    maxspeed  min/max_total_time
+
   // jumps in speed in real on esp32
   test.with_pars("f1", 1000, 4300, 10000, true, 4.5 - 0.2, 4.5 + 0.2, 0.5, true,
                  true);
@@ -285,6 +294,9 @@ int main() {
                  1.50 - 0.04 * nc, 0.1, true, false, true);
 
   test.with_pars("f25", 1000, 40, 0x7fffffff, true, 0.039, 0.041, 0.1);
+
+  // very short ramp. detected by esp32_hw_based tests seq_06.sh
+  test.with_pars("seq_06.sh", 54, 40, 1000000, false, 0.012, 0.018, 0.1);
 
   printf("TEST_02 PASSED\n");
   return 0;